+++ package.json
@@ -2,8 +2,9 @@
   "name": "@zondax/ledger-substrate",
   "author": "Zondax AG",
   "license": "Apache-2.0",
-  "version": "0.17.9",
+  "version": "0.35.2",
   "description": "TS / Node API for Substrate based apps running on Ledger Nano S/S+/X",
+  "bin": "./dist/cmd/cli.js",
   "main": "./dist/index.js",
   "typings": "./dist/index.d.ts",
   "types": "./dist/index.d.ts",
@@ -95,8 +96,7 @@
     "copy-files": "copyfiles -u 0 src/**/*.proto dist/",
     "test:integration": "yarn build && jest -t 'Integration'",
     "test:key-derivation": "yarn build && jest -t 'KeyDerivation'",
-    "supported": "ts-node src/index.ts supported",
-    "test:unit": "jest -c jest.config.unit.js",
+    "supported": "ts-node src/cmd/cli.ts supported",
     "linter": "eslint --ext .ts,.tsx,.js,.jsx --ignore-path .eslintignore . --max-warnings 0",
     "linter:fix": "yarn linter --fix",
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\""
